Pull upstream changes for the OpenAPI spec
Based on https://github.com/moby/moby/blob/69adaa894d829a6707427aef6aeb0a796b2fb385/docs/api/v1.41.yaml and https://github.com/moby/moby/blob/69adaa894d829a6707427aef6aeb0a796b2fb385/api/swagger.yaml

+++ api-model-v1-41/src/main/kotlin/de/gesellix/docker/remote/api/HostConfigAllOf.kt
@@ -192,7 +192,7 @@ data class HostConfigAllOf(
 
   /* Size of `/dev/shm` in bytes. If omitted, the system uses 64MB.  */
   @Json(name = "ShmSize")
-  var shmSize: kotlin.Int? = null,
+  var shmSize: kotlin.Long? = null,
 
   /* A list of kernel parameters (sysctls) to set in the container. For example:  ``` {\"net.ipv4.ip_forward\": \"1\"} ```  */
   @Json(name = "Sysctls")
